Create userController
The "userSignInPost" middleware verifies that all field values
entered by the user and sent to the server are correct. If so,
call "handleSchemaErrorMessages" to respond to field error messages;
if not, call "handleAuthenticate" to verify whether the user is
registered. If so, call "handleSignIn" to respond to the token;
if not, call "handleSignInErrorMessages" to respond to the login
error message.

The "userSignUpPost" middleware verifies that all field values
entered by the user and sent to the server are correct. If so,
call "handleSchemaErrorMessages" to respond to the field error
message; if not, call "handleSignUp" to hash the password and
register the user into the database, and then log in the user.