Create postController
The "postList" middleware handles getting all posts and author name.

The first middleware of "postDetail" calls "verifyParamId" to check
parameter ID, and the second middleware handles getting a specific
post and author name.

The first middleware of "postCreate" calls "verifyToken" to check
whether the user is logged in, the second middleware "verifyPermission"
checks the user is an administrator or author, the third middleware
"verifySchema" checks whether there are errors in each input field,
and the last middleware handles adding new post.

The "postUpdate" middleware is similar to the "postCreate" middleware.
The difference is that "verifyParamId" is added in the second
middleware position to check whether the parameter ID format is
correct, and the last middleware handles updating a specific post.

The "postDelete" middleware is similar to the "postUpdate" middleware.
The difference is that "verifySchema" is not added in the fourth
middleware position, and the last middleware handles deleting a
specific post.