Create commentController
The "commentList" middleware handles getting all comments for a
specific id and author name.

The first middleware of "commentCreate" calls "verifyToken" to check
whether the user is logged in, the second middleware "verifyPermission"
checks the user is an administrator or author, the third middleware
"verifySchema" checks whether there are errors in each input field,
and the last middleware handles adding new comment.

The "commentUpdate" middleware is similar to the "commentCreate"
middleware. The difference is that "verifyParamId" is added in the
second middleware position to check whether the parameter ID format is
correct, and the last middleware handles updating a specific comment.

The "commentDelete" middleware is similar to the "commentUpdate"
middleware. The difference is that "verifySchema" is not added in
the fourth middleware position, and the last middleware handles
deleting a specific comment.