Updated "verifyId" middleware
Renamed parameter to type.

The error response of the first middleware is changed to 400 status.

The error response of the second middleware changes to a 400 status,
and the error message also changes.

The Reply model is added to the last middleware. If the data is found,
the author ID of the data is set as a local variable.