Updated tokenExChange
Deleted conditional for rid variable in the first middleware.

The third and fourth middleware are merged into a new middleware
that stores the refreshToken document in the request variable.

The fourth middleware handles the token is expired or other errors.

The last middleware updates notBefore property on the refreshToken
of request variable instead of using the "findByIdAndUpdate" method.