Updated postUpdate
The verifyJSONSchema middleware adds unescape sanitizer to the title
property, a conditional to the mainImage property and a custom
validator to the content property to verify the word count.

The verifyJSONSchema middleware adds unescape sanitizer to the title
property, a conditional to the mainImage property and a custom
validator to the content property to verify the word count.

The fifth middleware uses the post document that stored by "verifyId"
middleware, to update post instead of the "findByIdAndUpdate" method.