feat: changed the replyCreate middleware
The third middleware finds a comment, and it has a specified reply id.

The last middleware pushes the new reply ID to the reply's parent comment's child array
when creating a reply to another reply.

And finding the new reply and populating the author's username and reply info,
then it returns a response with the created comment.