feat!: changed the replyUpdate middleware
Finding a specified reply will not populate author property in the third middleware.

The fourth middleware adds a reply variable to assign the value of reply author information.

The last middleware saves the new reply's content
and populating the author's username reply info, then it returns a response with the updated reply.